+++ tests/AvroConvertTests/CodecTests/Brotli/BrotliComponentTests.cs
@@ -89,7 +89,7 @@ namespace AvroConvertComponentTests.CodecTests.Brotli
             //Assert
             Assert.NotNull(result);
             Assert.NotNull(deserialized);
-            Assert.Equal(toSerialize.StringProperty, deserialized.StringProperty);
+            Assert.Equal(toSerialize.NullableStringProperty, deserialized.NullableStringProperty);
             Assert.Equal(toSerialize.NullableIntProperty, deserialized.NullableIntProperty);
         }
     }