+++ tests/AvroConvertTests/CodecTests/Snappy/SnappyComponentTests.cs
@@ -90,7 +90,7 @@ namespace AvroConvertComponentTests.CodecTests.Snappy
             //Assert
             Assert.NotNull(result);
             Assert.NotNull(deserialized);
-            Assert.Equal(toSerialize.StringProperty, deserialized.StringProperty);
+            Assert.Equal(toSerialize.NullableStringProperty, deserialized.NullableStringProperty);
             Assert.Equal(toSerialize.NullableIntProperty, deserialized.NullableIntProperty);
         }
     }