+++ tests/AvroConvertTests/DefaultSerializationDeserialization/AvroAttributeClassTests.cs
@@ -30,7 +30,7 @@ namespace AvroConvertComponentTests.DefaultSerializationDeserialization
             Assert.NotNull(result);
             Assert.NotNull(deserialized);
             Assert.Equal(toSerialize.NullableIntProperty, deserialized.favorite_number);
-            Assert.Equal(toSerialize.StringProperty, deserialized.name);
+            Assert.Equal(toSerialize.NullableStringProperty, deserialized.name);
             Assert.Equal(toSerialize.AndAnotherString, deserialized.favorite_color);
         }