+++ tests/AvroConvertTests/Headless/HeadlessComponentTests.cs
@@ -49,5 +49,34 @@ namespace AvroConvertComponentTests.Headless
             Assert.NotNull(deserialized);
             Assert.Equal(toSerialize.justSomeProperty, deserialized.justSomeProperty);
         }
+
+        [Fact]
+        public void Component_AvroAttributeClass_ResultIsEqualToInput()
+        {
+            //Arrange
+            AttributeClass toSerialize = new AttributeClass
+            {
+                AndAnotherString = "anotherString",
+                NullableIntProperty = 1,
+                NullableIntPropertyWithDefaultValue = null,
+                NullableStringProperty = "nullableString"
+            };
+            string schema = AvroConvert.GenerateSchema(typeof(AttributeClass));
+
+
+            //Act
+            var result = AvroConvert.SerializeHeadless(toSerialize, schema);
+
+            var deserialized = AvroConvert.DeserializeHeadless<AttributeClass>(result, schema);
+
+
+            //Assert
+            Assert.NotNull(result);
+            Assert.NotNull(deserialized);
+            Assert.Equal(toSerialize.NullableIntProperty, deserialized.NullableIntProperty);
+            Assert.Equal(toSerialize.AndAnotherString, deserialized.AndAnotherString);
+            Assert.Equal(toSerialize.NullableStringProperty, deserialized.NullableStringProperty);
+            Assert.Equal(2137, deserialized.NullableIntPropertyWithDefaultValue);
+        }
     }
 }